move EntityModifiedLogsServlet to ui.controller
